pyproject.toml
[project]
name = "sui-gas-pool-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Studio Mirai, LLC", email = "dev@sm.xyz" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.2",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "respx>=0.20.0",  # For mocking httpx requests
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

[project.scripts]
sui-gas-pool-sdk = "sui_gas_pool_sdk:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "python-dotenv>=1.0.1",
]
